+++ server-side-rendering/src/lib/oauth.ts
@@ -8,7 +8,7 @@ export function signInWithGithub() {
 	const url = new URL(`${PUBLIC_APPWRITE_ENDPOINT}/account/sessions/oauth2/github`);
 
 	url.searchParams.set('project', 'ssr');
-	url.searchParams.set('success', `${window.location.origin}/api/auth`);
+	url.searchParams.set('success', `${window.location.origin}/auth`);
 	url.searchParams.set('failure', `${window.location.origin}/`);
 	url.searchParams.set('token', `true`); // Crucial: Set token to true to get a auth token in the success URL
